Order printouts: Move prinouts to tab from toolbar
Refs SH-71
