+++ shuup/order_printouts/templates/shuup/order_printouts/admin/section.jinja
@@ -1,0 +1,21 @@
+<div class="row">
+    <div class="col-xs-12 col-sm-12 col-md-4">
+        <h2>{% trans %}Confirmation{% endtrans %}</h2>
+        <a href="{{ url('shuup_admin:printouts.confirmation_pdf', order.pk) }}"
+           class="btn btn-info btn-block" target="_blank">
+            {% trans %}Get Order Confirmation{% endtrans %}
+        </a>
+
+        {% set shipments = order.shipments.all() %}
+        {% if shipments|length %}
+            <hr>
+            <h2>{% trans %}Shipments{% endtrans %}</h2>
+            {% for shipment in shipments %}
+                <a href="{{ url('shuup_admin:printouts.delivery_pdf', shipment.pk) }}"
+                   class="btn btn-info btn-block" target="_blank">
+                    {% trans id=shipment.pk %}Get Delivery Slip (Shipment {{ id }}){% endtrans %}
+                </a>
+            {% endfor %}
+        {% endif %}
+    </div>
+</div>
